pyproject.toml
[project]
name = "denodo-statistics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "denodo-sqlalchemy>=1.0.6",
    "pyyaml>=6.0.2",
    "ruff>=0.8.2",
    "sqlalchemy>=1.4.54",
]
[tool.setuptools]
packages = ["automation", "utils"]
